Validated Q1 and Q2, started Q3
Used pandas to validate Q1 and Q2. Started Q3. I have the matrices ready for training. Not sure how dirty the data remains. Still need to fit models and validate results with my results from HW1. Might need to convert it to a class based method so that I can reliably re-use it.